Update tarball download links to the official URL
As discussed on the mailing list, Junio's release process is more than
just running a git archive command or equivalent. The official tarballs
contain a pregenerated configure script and adds a correct version file.

The link has been updated to point to the official URL for downloading
the git tarballs, as specified by the git [ANNOUNCE] messages on the
mailing list.

Signed-off-by: Dan Johnson <ComputerDruid@gmail.com>