donloads: show the installer/package release date
Yet Another Msygit Fix. Store the installer's release date as a
separate field, as it's not always the date that the version was
tagged.

This allows us to report the correct release date in the /download/win
page. Unfortunately the msysgit downloads aren't always uploaded to
github on the date that the installer has, so we need to parse it to
get the release date.