adjust amazon affiliate links
The Pro Git links to Amazon use Scott's personal affiliate
id, which puts Git the non-profit project in a murky area.
This site is not owned or managed by the Software Freedom
Conservancy's member project Git, but it's de facto status
as the official site makes things awkward. And anyway, it
would be nice to donate the affiliate revenue to the
project.

This patch replaces all of the affiliated Pro Git links with
ones that mention the Git/SFC affiliate id.