make irc channel descriptions more neutral
The intent of this site is to be about Git, not GitHub, and
it mostly does a good job of not mentioning GitHub at all.
One exception is the community page, which lists Freenode IRC
channels for #git and #github. Let's be more neutral and
note that #git is the right place for git advice, and
provider-specific channels can be used for advice on those
systems.