Created retina versions of all graphical assets.
This commit includes @2x versions of all graphical assets on the site except for the images in the Pro Git book. This includes the favicon, all icons, the branching illustration on the front page, the OS monitors, screenshots of the GUI apps, book covers, etc. I also recreated the images used in the About section.

I created a new Sass mixin for showing @2x background images in the CSS. This will render the regular image unless it meets the media query criteria for Retina display.

I also ran all of the PNGs through a compressor (tinypng.org) and reduced their sizes quite a bit.