add contributors page
The old version of git-scm.com featured a list of
contributors on the "about" page, and the git(1) manpage
linked to it. That list is no longer there, but it would be
nice for the manpage to be able to link to something.

This patch recreates the contributor list from the old site
as "/contributors". It isn't linked from anywhere else on
the site; it might make sense to link it from the "/about"
or from "/community".

It builds with a local index using git-shortlog, but the
update to index.rake is questionable, for two reasons:

  1. It makes O(# of contributors) Octokit calls to get the
     name of each contributor, which is slow.

  2. The Octokit calls are pulling the contributors list
     from GitHub's copy of gitster/git. However, those
     values do not match very well with git-shortlog. For
     example, Linus is much lower than he should be. This
     may not be a good list to pull from.