add new Packt Publishing git book
This is a new git book coming out in January 2013. I haven't
seen it yet, so I don't know about the quality, but I think
we are better to be inclusive in this list. There are not
that many git books, and unless one of them is known to be
really terrible or obsolete, I don't see any reason not to
link to them all.