Initial preparations for i18n
This is the initial batch of preparation for readying the site to be i18ned. I'm using rails' standard way to i18n applications, and have started to extract the static content from erb and haml templates and move it to yaml files.

I've tried to organize the yaml files in the best way possible to resemble the actual view hierarchy and be logical enough to make it easy to add new translations.

Not all files have been processed. I wanted to get this out early to get some feedback on how the files are organized and if this is the best way to move forward with the internationalization of git-scm.com