Pointing Mac downloads to Sourceforge instead of Google Code
The downloads for the Mac installers moved to Sourceforge, so I've
updated the downloads.rake and the downloads controller and haml file
to point to the new location. Note that SF doesn't allow deep linking,
so the iframe trick doesn't work here (you'll always need to manually
download it, the browser won't force a download).