about:free-and-open-source: do not paraphrase GPL
Trying to summarize or paraphrase the GPL in laymans' terms may have
meant well, but it will invite legal issues later.  A carelessly
written paragraph that appears on the official home page of the Git
project can create a loophole to avoid license compliance, when there
is a contradiction between the loosely written version and the license
text itself, the latter of which must be the authoritative and single
source of truth.

We (the Git committee) were advised by a Software Freedom Conservancy
lawer to state the reason why the project chose to use GPL, namely, to
make sure our software will stay free, instead.

Signed-off-by: Junio C Hamano <gitster@pobox.com>