layouts: only use "site" layout for site#index
The /sfc page is rendered with the masthead and no navbar,
just like the main page. This makes it hard to read, and
hard to navigate away from. This is likely an unintended
consequence of c575cbd.

Let's only use that layout for the index page. All of the
other pages handled by the site controller (/sfc,
/courses/svn, and the search results page) probably want to
be treated more like "regular" pages.

Fixes #388.