Add mime-type to ebook download links
The download links on the site are untyped for the ebooks. As reported
in progit/progit#796 this prevents from downloading the ebooks directly
on the reader's tablet.

This change is only compatible with html5 and may not work on old
browser. A better alternative would be to set up the mime types for epub
and mobi in the web server configuration files.