downloads: support 32-bit/64-bit windows links
With the switch to the *MSYS2* backend for the *Git for Windows* project
the opportunity arose to support both `32-bit` and `64-bit` builds. In
additon to those `32-bit` and `64-bit` builds *Git for Windows* now ships
with portable releases as well. So it makes sense to display all those
links on the download page. Since this behavior differs from the `mac`
platform would be best to introduce a new download page for the `windows`
platform.

The naming convention for *Git for Windows* builds changed too, so the
`rake downloads` task had to be adapted. It now parses the *Git for
Windows* release file properly and also extracts the `bitness` and the
portable information of the release file. That `bitness` and the portable
information is then appended to the platform when storing the release
information in the database.

To provide an automatic download for the latest *Git for Windows* release
in the correct `bitness` parsing of the user agent in the download
controller was added. The view was updated to show the current download
and then list the other options as well.

To associate the latest Windows release (which might have 4 digits) to the
corresponding Git version, we look only at the first 3 digits (which match
the Git version). Thanks to the order in which the GitHub API delivers the
releases, this means that the last one wins, as it should be. Example: if
2.5.0 was uploaded, and later 2.5.0.2 was uploaded, then 2.5.0.2 is added
as a `Download` *later* than 2.5.0, and associated with the Git version
2.5.0. So even if it might not be completely obvious from a quick glance,
the logic is sound.

[jes: rebased to current upstream; accounted for 3-digit versions,
commented on the 4 -> 3 digits reduction to match up Git for Windows to
Git versions]

Signed-off-by: 마누엘 <nalla@hamal.uberspace.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>