genbook2: find &amp;gt; inside split <code> blocks
Many of these cases are in blocks like this (lines split for
readability):

    <code class="p">&amp;</code>
    <code class="n">gt</code>
    <code class="p">;</code>

with some variations (sometimes no code block in the middle
one, sometimes other classes, sometimes extra characters at
the beginning of the leading block or the end of the
trailing one).

This is a regex that should catch all of them, while
remaining sufficiently strict that we won't hit any false
positives in practice.