update v2 book images from progit/progit2
This is basically just a mass-import of every png in
progit2's book/ directory, which appears to be what was done
originally by @schacon when the book was initially imported.

The main goal is to pick up any fixes (such as the one
in https://github.com/progit/progit2/pull/178), but it's
hard to tell which images have changed content. In
https://github.com/progit/progit2/pull/424, all of the
images were png-optimized, so just looking at the tree-diff
doesn't say much. Literally every file image was touched
there.

Still, these are the images going into the current builds of
the ebooks, so one assumes they are what we should be
showing.