Gemfile: bump to latest asciidoctor
This has tabsize fixes from asciidoctor/asciidoctor@2893d1666060
that make some of the mixed tab/space diagrams line up
properly (e.g., the first one in git-merge-base).

Actually, that commit is in 1.5.3, but it seems like it
should not hurt to get the latest round of bugfixes, too.

Fixes #680 and #847.