doc_helper: handle topic guides in man() linker
We strip `git-` from the front of command names in the link
text. But that doesn't handle topic guides like
`gitrevisions`.  Let's use a regex to handle both at once.

While we're at it, we can tighten this up to strip only from
the front of the command (no commands actually have "git" in
later in their name, but it is an accident waiting to
happen).