drop asciidoctor templates
These originally came from 5a5edec (Add Asciidoc
parsing/rendering, 2012-04-03), but these days asciidoctor
generates almost the exact same content. I verified this by
diffing the results of running asciidoctor on all of the
v2.10.2 manpages, with and without the templates, and
diffing the result.

The only differences were:

  1. Our sectanchors option now works! Yay!

  2. In user-manual.html, the top-level sections are all
     level "0", which renders slightly differently. Our
     templates put it in a `<div class="sect0">`, but
     asciidoctor just adds `class="sect0"` to the `<h1>`
     tag.

     This shouldn't matter either way, as we do not style
     `sect0` at all (and it doesn't affect the manpages at
     all; just the user manual). But arguably the
     asciidoctor output is better anyway.

So let's stop using our own templates. It's less code of
ours to carry around, it unbreaks sectanchors, and it lets
us benefit from future bugfixes and features in
asciidoctor's upstream html output.

Note that the option name was mispelled in
`local_index.rake`, meaning that the templates were
_already_ not being used in development environments. Only
the live site, which uses `preindex`, did so.