preindex: re-handle missing include files
The recent change in #918 to handle built category files
lost the check for when an included file is not present (we
need to see that content_file is nil, and avoid looking at
its sha).

While we're fixing that, let's also bump the examination of
content_file down into the else block, too. if we have an
entry in categories, then we do not use the result at all.