book: mention Greek translation
I've listed it as a partial, since it's not done. I'm
unclear on the distinction between "partial" and "started",
but this seems to be about as far along as some of the other
partials.

I pulled the native word for "Greek" from Wikipedia.
Hopefully it's right, but I've asked a native speaker to
double-check.