tasks: add reset_book2 task
When we update the build recipe, all of the book codes need
rebuilt. We have GENLANG, but there's no "rebuild all" flag.
That would be easy to add, but it isn't very robust.
Building the books takes long enough that a "heroku run
rake" job dies partway through, and we have no idea which
were rebuilt and which weren't.

Instead, let's reset ebook html flag (which is hackishly used to
store the tip sha1 of what has been built). Then
"remote_genbook2" can be triggered as necessary to do the
actual builds.