books: fix historical redirects
When we see "/book/4_git_treeishes", we route it via a
redirect table to "/book/en/Git-Tools-Revision-Selection".
However that URL is generated by feeding "en/..." to the
book_lang_path() helper from the routes table, which quotes
the "/", breaking the link.

We can work around this by splitting the two parts and
feeding them to the "book/:lang/:slug" route instead.