In preparation for using Hugo, move the "blog has moved" page
We will use GitHub Pages to serve https://git-scm.com/, as a Hugo
site. To that end, move the "blog has moved" page to a new location.

In Hugo's world, there are two ways `index.html` files can be provided:
as `index.html` or as `_index.html` file. The former is treated as a
"single" page while the latter is treated as a "list" page. Since we
want https://git-scm.com/blog to show that page, the `_index.html` is
used here.

Signed-off-by: Matt Burke <spraints@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>