Create a GitHub Issue Template.
People *constantly* open issues on the git/git-scm.com repository that
aren't actually relevant to the git-scm.com website. This introduces an
issue template which explicitly points people to the appropriate issue
tracker for the Git, Git for Windows, and Pro Git projects, so that anyone
who goes to open an issue on the git/git-scm.com issue tracker that
isn't about the website gets told where to go.