point elasticsearch at $BONSAI_URL
I'm not sure why the $SEARCH_INDEX_URL variable is used. It
does provide a layer of indirection, and the existing site
has both Bonsai and Searchify set up. So maybe it was for
switching between them.

These days we're just using Bonsai, so let's point it at the
variable that is automatically set up by the Heroku add-on.