app/views/community: clarify subscription and archive are two separate things
The original seounded as if it wanted to teach how to subscribe to
the public-inbox archive, which does not make much sense.  Separate
the two and then explain the benefit of subscribing a bit better by
taking suggestion made by Ævar on the mailing list (<878tkjk7m7.fsf@gmail.com>).

Signed-off-by: Junio C Hamano <gitster@pobox.com>