tasks/index: fix link:technical regex
The "\1" and "\2" here are regex replacements, but the
double-quotes mean they get interpolated to the literal
bytes "1" and "2". We could fix this with "\\1", but it's
simpler still to just use a single-quoted string.

We can also get rid of the backslash-escaping of the
brackets. They're not needed on the replacement side of the
regex (and in fact we must get rid of them to not have them
show in the non-interpolated string).