tasks/local_index: set asciidoctor `doctype` to `book`
This matches the change done to the preindex task in
5f7ca46e95f0e5506baa025cad8248309ba8d2e0. This shouldn't
affect most documents we build, since asciidoctor assumes
them to be manpages based on their content. But it does shut
up a bunch of "section 0" warnings when building
user-manual.txt.

I can't detect any visual change to the results, so this is
at least (hopefully) not regressing anything.