Factorize the index rake task
The index task is in charge of gathering the documentation from the
git project and injecting it into database. Until now there was two
distinct files for retrieving the git project, one, `index.rake` which
allowed to get the files'content from github, and another one,
`local_index.rake` which allowed to get the content from a git
repository on the local disk. Both share the same logic in building
the pages and only differ on how to access them.

The purpose of this commit is too reunit the common logic. From now
on, only one logic of page building is provided.