downloads: reword the ubuntu ppa note
I wanted to get rid of "from the developers" here, because
the PPA isn't package by Git developers. It's done by
third-party Ubuntu folks. I have no reason to think they
won't do a good job of keeping up to date, but I think it
makes sense to keep it clear that it's a third party.