index.rake: uncomment loop-next for found tags
If we've already got a version for a particular tag, we
don't need to reimport it. However, in the transition to the
shared local/preindex code, this line got commented out
(probably because it was useful for debugging and got left
in).

You can always override it setting the RERUN environment
variable.