drop /admin route
This just shows a list of the downloads pulled down from the
`download` task. Potentially useful for remote-admin, I
guess, but it falls far short of what's needed for actual
debugging (and I can't think of a single time when it would
have been helpful).

It also seems to cause rendering exceptions on the
production site when spam requests look at it.

Let's just drop it as ancient and not-useful.