Revert "Merge pull request #1047 from git/raygun"
I added Raygun support to get a handle on the exceptions we
were seeing due to timeouts in #1045. That's now reverted,
so the primary reason is gone.

Furthermore, the Intro plan only allows 500 exceptions per
month, and we exceeded that pretty quickly. Most of these
are just noise, so we may want to consider a service like
this after cleaning them up. But in the meantime, it's not
accomplishing anything.

The noisy exceptions seem to mostly be related to book
languages with broken links or incomplete translations. They
fall mainly into two buckets:

  - dereferencing nil for various items. E.g., hitting
    https://git-scm.com/book/sr/v1/%D0%9F%D0%BE%D1%87%D0%B5%D1%82%D0%B0%D0%BA-Getting-Help
    returns the 500 error page.

  - ActionController::BadRequest. E.g., hitting https://git-scm.com/book/ko/Git-%EC%EF%BF%BD%EF%BF%BD%EB%B2%EF%BF%BD-GitWeb
    returns no output