index.rake: order the build of manpages in ascending versions
Using a simple alphabetical sort did not lead to correctly ordering
the versions to build. It is not known if it has some effects on the
issue of presenting the versions out of order on the website, but it
is safer anyway to stick to correct ordering for insertions into the
database.