Add latest version available information on download 'monitor'
The information depends on the platform the user is using to access the site.
For windows, since on the DB there are several windows versions
(windows32, windows64, windows32Portable, windows64Portable)
I decided to go with windows32. Portable versions are almost never the
ones we want. windows32 can always be installed, even on windows64, so
decided to go for a conservative approach.