bump rails to 4.2.10
This has some fixes which are required to run with the newer
ruby we just bumped to in the last commit. Note that I
manually bumped the version in Gemfile, but then did a
"bundle update" to pick up the updated dependent packages.

This means we _also_ updated any other gems which had new
versions. This is probably a good thing, as many of them
likely have security fixes.

Most notably, this brings us to asciidoctor v1.5.6.1, which
is full of bugfixes and improvements. This should in theory
be a good thing, but it's possible there may be regressions
in the manpages. These won't show up immediately (since we
only run asciidoctor at import time), but may crop up as we
import new versions.