guis: bump "X GUIs shown below" to its own line
As we add more selection buttons, this line is liable to get
wrapped if you have the text zoomed at all. Let's just put
it on its own line, which looks equally good (actually, I
like it better even when it's not wrapped).

We can revert the padding tweak done in 44b9431 (Add android
and ios GUI sections, 2017-11-10), which was purely about
avoiding that wrapping.