remote_genbook2: handle repos with no releases
If there are no releases for a translation, then we get an
Octokit::NotFound exception when asking for the release
page. That causes us to skip our "book.save" call, meaning
we never actually update the database with the newly
imported translation (and we just try it fruitlessly in
every run).

We can fix that by rescuing the exception (the code was
already designed to handle this case by looking for a nil
release list).