session.js: disable gapi_location option
We don't use session.js for anything except the os
detection. So there's no need for us to actually turn on the
location-finding magic which causes the browser to hit
google's API.

I suspect that this could be done with a lot simpler
javascript than all of session.js (it looks like it's just
pattern matching the user-agent etc anyway, which we could
do inline in site.js). But I'll leave that for now, making
this a pure cleanup with little chance of regression.