drop unnecessary bits from package-lock.json
This is actually the result I get running `npm install`. It
seems like there are a bunch of unnecessary dependencies in
the package-lock file, some of which have security
vulnerabilities. These don't overly matter, since we use
node only for the stylelint check, but if we're not using
them, it's better to be tidy.

Note that npm also seems to want to use "^" rules instead of
"~" in the actual package.json.