redirect /book/commands to /docs
This currently generates a poorly formatted list of commands
with no actual links to the manpage. It was originally
referenced as part of a command index in a mock-up that went
away in the previous commit.

Probably nobody links to it, but we can easily just redirect
it to our good-looking link-filled command reference page.

Note that this lets us drop the redundant list of commands
in lib/constants.rb.