drop UPDATE_TOKEN code
Once upon a time, the idea was that you could update the
site with new book versions by posting to a special endpoint
with a secret token.

But these days we fetch the book material directly to the
rails environment by running a periodic rake task, and then
update the database directly. Let's drop this unused code.