drop blog content
There hasn't been a blog post in 7 years, and there are no
plans to revive it. The existing posts generally fall into
one of two categories:

  - news about the pro git book

  - technical notes that weren't covered in the 1st edition
    of pro git

The former are no longer interesting at all. And the latter
had their content folded into the second edition.

So let's drop this content, as we sometimes get issues
raised about its accuracy and formatting. We want to
redirect people to the actually-maintained book content.

We'll put a short note in the /blog route to help anybody
who follows old links to these posts.