Turn external link tutorials subsections 2 columns each
Om external links, the Tutorials section has 2 subsections. One is using the left column,
while the other is using the right column. Since the number of items of each subsection
is now very different, the UI is not very simmetric and has a lot of white space.

This change makes each of the subsection to show it's elements on two columns, solving
these problems