book2: destroy a book edition before creating it
If a book deletes chapters or sections (or even adds a new
chapter in the middle) we can be left with old sections left
in the database. For instance, if we add a new chapter 4
with three sections, then if the old chapter 4 had five
sections, we have two leftover bits of cruft that will
appear in the table of contents.

Let's start each book import with a clean slate. Because
we're working in a transaction, we can freely destroy this
and then re-add the new content without any disruption to
people viewing the site.