Replace Tire gem by newer elasticsearch gem
This required some code changes in the point of contact with the client
but nothing big.

A few TODOs were added as i was navigating the code, for instance:

- There's a 'meta' property of the search results being populated,
but doesn't seem to be used. Need to confirm

- The 'type' property of ES documents is looking at the class name, which
is not very OOP. We could easily change it. But this takes me to the last
point

- We are using ES 2.4.0, while the latest version is 6.X, in which
the 'type' property is deprecated and marked to removal. There are also
some query format changes, but since we are using basic queries, it should
be fairly easy (I started testing this work with a 6.X version).
If we are going to ES update, the 'types' property is going to be
removed anyway: https://www.elastic.co/guide/en/elasticsearch/reference/6.0/removal-of-types.html