index.rake: rename expand() to un-break preindex target
The local expand!() function was renamed in 4179f33 (resolve relative
path resolution in manpages includes, 2019-03-05) to just expand().
That causes a bad interaction with one of our gems, which looks like
this when running `rake preindex`:

    ArgumentError: wrong number of arguments (given 1, expected 4)
    .../git-scm.com/lib/tasks/index.rake:92:in `expand'
    .../.rvm/gems/ruby-2.4.2/gems/sawyer-0.8.1/lib/sawyer/agent.rb:134:in `expand_url'
    .../.rvm/gems/ruby-2.4.2/gems/sawyer-0.8.1/lib/sawyer/agent.rb:92:in `call'
    .../.rvm/gems/ruby-2.4.2/gems/octokit-4.8.0/lib/octokit/connection.rb:156:in `request'
    .../.rvm/gems/ruby-2.4.2/gems/octokit-4.8.0/lib/octokit/connection.rb:19:in `get'
    .../.rvm/gems/ruby-2.4.2/gems/octokit-4.8.0/lib/octokit/client/objects.rb:62:in `blob'
    .../git-scm.com/lib/tasks/index.rake:160:in `block (2 levels) in <top (required)>'
    [...]

There's likely a better way of declaring it in a more local way, but I
won't pretend to know enough Ruby to say what it is. In the name of
expediency, let's just rename the function for now to unbreak things.