actually apply manpage stylesheet
Our man-pages.scss file looks for a .man-page class, but we do not
actually generate any html with that class. It's unclear to me if
something changed in asciidoc, or perhaps in our rendering code.

At any rate, none of these styles were being applied. Let's look for
"#documentation #main" instead, which seems to match the generated html.