manpages: fix links when in version or translation
The links need to be rewritten to remove the extra path containing the
requested version or language, while still retaining the anchor.

For version, we cannot know before hand which version is presently
viewed, so we just drop to the latest version.

For language, we can be smarter and direct to content translated in
same language.

fixes #1382