/community: mention MyFirstContribution tutorial
We build and host the "My First Contribution" tutorial since commit
44bf766 (index.rake: build MyFirstContribution tutorial, 2019-08-14),
but the tutorial is not featured anywhere on the site or in the man
pages so it can only be viewed if you know it exists!

Make it easier to discover by mentioning it in the "Contributing to Git"
section of the "Community" page.