re-add conservancy banner
It's end-of-year donation time again, so let's point visitors to
Conservancy's donation page.

This is effectively a revert of aa0c5d0 (drop conservancy donation
banner, 2019-01-09), but with the year in the dismissal tag bumped to
2019.

There's actually a pretty cute ascii-art banner they're providing at
https://sfconservancy.org/img/banners/2019-member-projects-banner.png,
but I had trouble making it look good on the site (it's quite tall if
you let it take the whole width).