ARCHITECTURE.md: update for Cloudflare teams
Cloudflare now supports shared ownership of domains, and we use this
feature to administrate both 'git-scm.org' and 'git-scm.com'.

Since our adoption of this feature, the 'DNS' section of the
architecture document is out of date, and instead refers to the
old-style single-account setup.

Update the documentation here to reflect the current architecture, so
that others in the future can volunteer to help if they wish to do so.