package.json: bump stylelint versions
Dependabot is nagging about a bunch of out-of-date npm packages with
vulnerabilities. These shouldn't be a big deal for us because we don't
run any node code in production. But it shouldn't hurt (and we're
grossly out of date anyway).

I had to manually include @primer/css for the result to actually work
(I'm not sure of the details, but this was some guess-and-check running
"grunt" after the update).