downloads: link to latest tarball version for Linux
The kernel.org directory for tarball is rather long these days. Let's
mention the latest version with a direct link to the tar.gz file (just
like I recenlty did for the macos page).

We could also link to the .xz, which is slightly smaller, but I worry
that users may xz-utils installed.