lib/tasks/index.rake: appease Rubocop
Rubocup prefers that we use 'Enumerable#reverse_each' instead of using
each method separately. This improves performance by not requiring
'#each' to slurp up the results of reversing the list before enumerating
it.

Switch out one for the other for a slight performance win, but mostly to
get Rubocop happy again.