robots.txt: drop comment lines
These don't do anything, and we have no need to document to ourselves or
to others what the format of a robots.txt file is. We _could_ delete it
entirely, but it's probably better to indicate that we do know about the
file, and just choose not to have any entries.