css: color code blocks within links as links
We normally color code blocks as $orange. But then if you do the
somewhat obvious thing of trying to put a code block in the link text:

  Try the <a href="/docs/git-foo"><code>git-foo</code></a>

it's hard to immediately see that it's actually a link. Let's color
these cases like a normal link.

I have a feeling there's a more elegant CSS solution (possibly using
!important on the link coloring?), but this seems to get the job done,
and the use of scss means we don't have to repeat the actual color
codes.