community: avoid code blocks for command names, etc
Our CSS specifies a color for <code> elements ($orange), which makes it
hard to see a linkified command name (because it's not the usual
$link-color). Let's just use a span with the "fixed" class here, which
is what other parts of this page do (they also use "highlight", but to
my eye it makes these cases look worse).

Another option would be to tweak the CSS to color linked code blocks
properly, but I think it looks a bit nicer in this context with just the
fixed-width font.