tasks/index: prefer "HEAD" over "master"
When there are no tags, we look for documentation in the hard-coded
master branch. This is unlikely for git.git, but presumably triggers for
translation repos. We should be more flexible and just use HEAD in this
case, which will do the right thing if the translation repos switch
their default branch.

Note that we don't to do anything clever here like we did for the
"book2" case. We're already using the commit API for the remote index,
so we're free to just look up the name "HEAD". And likewise the local
case is just running Git commands, so that name is fine.