tasks/book2: fix race with HEAD update
When we index book content, we use the sha1 of the HEAD commit as a
cache key to see whether anything has been updated. But we fetch the
tree content by asking for HEAD separately. It's unlikely, but the
repository could be updated between the two, meaning we'd store a
mis-matched cache key (and because we get the tree first, we'd be stuck
with the old data under the new key, and thus wouldn't even notice on
the next run).

We can fix this by getting the commit first, and then using the tree
sha1 it mentions, which is atomic. This is easy now that we're grabbing
the whole commit (and this also matches how we do documentation
indexing).