tasks/book2: fix bogus sha-checking
The recent change to use octokit.commit() to get the book data had one
mistake in it: we have to pull the commit sha directly from the
response (r.sha), _not_ from the commit element (r.commit.sha). But we
_do_ have to use the commit element to get to the tree sha.

As a result, we ended up with a blank sha field, meaning the first run
rebuilt everything, but subsequent runs would never rebuild. Oops.