FIX: hover links: only barely overlap content
Currently, the links for sections which appear on hover substantially overlap the content which they are placed to the left of. An overlap is necessary, because it prevents the link from disappearing as the user moves the pointer from the content to the link. However, too large of an overlap results in the user being unable to select the content for which the link is appearing. This is particularly undesirable for command arguments. This changes the overlap to be 0.01em, which brief testing indicated was sufficient to accomplish both goals.