css: adjust horizontal alignment of section links
This keeps the "only a little bit of overlap" behavior added by 332504fa
(FIX: hover links: only barely overlap content, 2020-10-23), but without
making any visual changes to where we see the section markers.

It may be that we want to move them around a bit, but let's keep this PR
focused on the functionality change, without any visual change.

The solution here was provided by @makyen.