swap out API_USER/API_PASS for GITHUB_API_TOKEN
The use of bare username/password combos to do API calls has been
dropped by GitHub, as of 2020-11-13:

  https://developer.github.com/changes/2020-02-14-deprecating-oauth-auth-endpoint/

Let's switch to expecting a token in the rake jobs that update the book
and manpage content. The PAT can be very limited, as it only needs to
access public repos (the only reason we need to auth at all is to
overcome API rate limiting).