assets: make sprockets manifest
This is necessary under newer versions of Rails in order to actually
compile stylesheets, etc.

It also seems to complain about some of the dependencies in the book
stylesheet which come from other sheets. Adding includes seems to make
it work. I'm not sure why this worked in the older versions.