stop using image_alt() function
This function auto-generates an alt tag based on the filename, but it
was deprecated because screen readers generally prefer to see a blank
alt tag than a possibly low-quality one.

We can drop this whole block, since it's just a fallback for
options[:alt] (and if that exists, we'd already just use it).