test: drop doc perf test
I can't get this to run on modern ruby/rails. The error looks like this:

  .../gems/minitest-5.14.2/lib/minitest.rb:1030:in `run_one_method': BrowsingDocTest#run _must_ return a Result (RuntimeError)

Looks like an incompatibility between minitest and rails-perftest,
according to:

  https://github.com/rails/rails-perftest/issues/43

There's even a patch which was merged:

  https://github.com/rails/rails-perftest/pull/44

but it looks like never released. I'm not sure I care enough about this
test to wrestle with it further. Comparing perf between versions might
be interesting, but fundamentally we can't _fail_ this in the automated
testing without some kind of baseline. So I don't think it brings much
value to the regression tests.